refactor(admin): create information tab
