+++ admin/app/components/users/user-detail-personal-information.hbs
@@ -1,10 +1,3 @@
-<section class="page-section">
-  <Users::UserDetailPersonalInformation::UserOverview
-    @user={{@user}}
-    @toggleDisplayAnonymizeModal={{this.toggleDisplayAnonymizeModal}}
-  />
-</section>
-
 <section class="page-section user-authentication-method">
   <Users::UserDetailPersonalInformation::AuthenticationMethod
     @user={{@user}}
@@ -31,14 +24,6 @@
     @removeParticipation={{@removeParticipation}}
   />
 </section>
-
-<ConfirmPopup
-  @message="Êtes-vous sûr de vouloir anonymiser cet utilisateur ? Ceci n’est pas réversible."
-  @confirm={{this.anonymizeUser}}
-  @cancel={{this.toggleDisplayAnonymizeModal}}
-  @submitButtonType="danger"
-  @show={{this.displayAnonymizeModal}}
-/>
 
 <ConfirmPopup
   @message="Êtes-vous sûr de vouloir dissocier ce prescrit ?"
