+++ admin/app/controllers/authenticated/users/get/information.js
@@ -2,7 +2,7 @@ import Controller from '@ember/controller';
 import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 
-export default class GetController extends Controller {
+export default class UserInformationController extends Controller {
   @service notifications;
 
   AUTHENTICATION_METHODS = {
