+++ admin/app/routes/authenticated/users/get/information.js
@@ -1,0 +1,7 @@
+import Route from '@ember/routing/route';
+
+export default class UserProfileRoute extends Route {
+  async model() {
+    return this.modelFor('authenticated.users.get');
+  }
+}
