+++ admin/app/templates/authenticated/users/get/information.hbs
@@ -1,0 +1,7 @@
+<Users::UserDetailPersonalInformation
+  @user={{@model}}
+  @removeAuthenticationMethod={{this.removeAuthenticationMethod}}
+  @addPixAuthenticationMethod={{this.addPixAuthenticationMethod}}
+  @reassignAuthenticationMethod={{this.reassignAuthenticationMethod}}
+  @removeParticipation={{this.removeParticipation}}
+/>
