+++ admin/tests/unit/components/users/user-overview_test.js
@@ -1,15 +1,15 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
-import createGlimmerComponent from '../../../../helpers/create-glimmer-component';
+import createGlimmerComponent from '../../../helpers/create-glimmer-component';
 import ENV from 'pix-admin/config/environment';
 
-module('Unit | Component | users | user-detail-personal-information', function (hooks) {
+module('Unit | Component | users | user-overview', function (hooks) {
   setupTest(hooks);
 
   module('#externalURL', function () {
     test('it should generate dashboard URL based on environment and object', async function (assert) {
       // given
-      const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
+      const component = createGlimmerComponent('component:users/user-overview');
 
       const args = {
         user: {
@@ -36,7 +36,7 @@ module('Unit | Component | users | user-detail-personal-information', function (
     module('when user already has an email', function () {
       test('it should allow email modification', async function (assert) {
         // given
-        const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
+        const component = createGlimmerComponent('component:users/user-overview');
         const user = { email: 'lisa@example.net', firstName: 'Lisa', lastName: 'Dupont' };
         component.args.user = user;
 
@@ -48,7 +48,7 @@ module('Unit | Component | users | user-detail-personal-information', function (
     module('when user has an username', function () {
       test('it should also allow email modification', async function (assert) {
         // given
-        const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
+        const component = createGlimmerComponent('component:users/user-overview');
         const user = { username: 'lisa.dupont', firstName: 'Lisa', lastName: 'Dupont' };
         component.args.user = user;
 
@@ -60,7 +60,7 @@ module('Unit | Component | users | user-detail-personal-information', function (
     module('when user has neither a username nor an email', function () {
       test('it should not allow email modification', async function (assert) {
         // given
-        const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
+        const component = createGlimmerComponent('component:users/user-overview');
         const user = { firstName: 'Lisa', lastName: 'Dupont' };
         component.args.user = user;
